Revert "ユーザーIDと販売者IDの型をLongからIntに変更"
This reverts commit 740130f3751bde8ea56a1016b4decdb3d78739af.

+++ src/main/kotlin/com/example/myapp/dto/OrderDto.kt
@@ -5,7 +5,7 @@ import java.math.BigDecimal
 
 // 注文作成リクエスト
 data class CreateOrderRequest(
-    val buyerId: Int,
+    val buyerId: Long,
     val productId: Long,
     val shippingAddress: String?,
     val shippingPostalCode: String?,
